enable ecar name agnosticity
